+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Every page now has a unique page title
+- Add transfer project statistics to statistics page
 
 ### Changed